refactor: cambio de ubicacion interfaz seccion 3
