+++ src/app/Core/models/icon-if.ts
@@ -1,12 +1,3 @@
-export interface ProjectIF {
-
-    title:string,
-    description:string,
-    tecnologies:Array<IconIF>,
-    image?: string
-
-}
-
 export interface IconIF {
     name:String
 }
