+++ src/app/Main/second-page/project-card/project-card.component.css
