+++ src/app/Main/second-page/projects-models/project-if.ts
@@ -1,0 +1,10 @@
+import { IconIF } from "../../../Core/models/icon-if";
+
+export interface ProjectIF {
+
+    title:string,
+    description:string,
+    tecnologies:Array<IconIF>,
+    image?: string
+
+}
